debug

+++ pyramid-mos/pLSTM_decoder.py
@@ -49,7 +49,7 @@ class Decoder(nn.Module):
 
 		if(train):
 			print("text:", text.shape)
-			text = torch.transpose(text, 1, -1)
+			text = torch.transpose(text, 1, -1).long()
 			#text = torch.transpose(text, 0, 1)
 			print("text:", text.shape)
 			max_len=text.shape[1]
